fixed size_t, ptrdiff_t on 64b

+++ gostdc.h
@@ -245,10 +245,12 @@ int setjmp(jmp_buf);
 
 //typedef void* ptrdiff_t;
 //typedef unsigned long long int ptrdiff_t;   // 64b expected
-typedef unsigned int ptrdiff_t; // 32b; will it be ok on 64b PCs?
+//typedef unsigned int ptrdiff_t; // 32b; will it be ok on 64b PCs?
+typedef uintptr ptrdiff_t;
 
 //typedef unsigned long long int size_t;      // 64b expected
-typedef unsigned int size_t; // 32b, to match size described in Lua header for chunks compiled by lua.exe
+//typedef unsigned int size_t; // 32b, to match size described in Lua header for chunks compiled by lua.exe
+typedef uintptr size_t; // NOTE: on 64b machines/OSs, compiled chunks will be incompatible with 32b
 
 #endif
